split coordination.py module into plotly.py and helpers.py
