+++ pymatviz/coordination/__init__.py
@@ -1,0 +1,3 @@
+"""Coordination number analysis and visualization."""
+
+from pymatviz.coordination.plotly import coordination_hist, coordination_vs_cutoff_line
